Darken action icons from #6e6e6e to #404040
These icons were originally made as a compromise to be visible both in the
dark themes and the light one. This resulted in them being functional but
not looking good in any of the themes.

We're now moving towards different icon sets for dark and light themes, this commit
darkens the icons in the standard set to adapt better to the light theme.

See https://github.com/linuxmint/mint-themes/issues/197