Remove input-keyboard
This icon is used in very different contexts and so it can't be properly
themed:

- In Fcitx it's used to show a keyboard in the status tray.. here it
should look like other status icons, preferably round and colorful.

- In MATE it's used to launch the keyboard preferences app.. here it
should look like an app, preferably like other apps, with a frame around
it.

- In Cinnamon it used to be used for the virtual keyboard launcher, here
again it had to look like an app.

- In various app windows it's probably used as a dialog icon to represent
a fullcolor keyboard device. Here it should look good when big and without
a frame.